+++ src/PiPedalModel.cpp
@@ -82,7 +82,7 @@ PiPedalModel::PiPedalModel()
     this->updater = Updater::Create();
     this->updater->Start();
     this->currentUpdateStatus = updater->GetCurrentStatus();
-    this->pedalboard = Pedalboard::MakeDefault();
+    this->pedalboard = Pedalboard::MakeDefault(Pedalboard::PedalboardType::MainPedalboard);
 #if JACK_HOST
     this->jackServerSettings = this->storage.GetJackServerSettings(); // to get onboarding flag.
     this->jackServerSettings.ReadJackDaemonConfiguration();
@@ -228,7 +228,14 @@ void PiPedalModel::Init(const PiPedalConfiguration &configuration)
     this->jackConfiguration = this->jackConfiguration.JackInitialize();
 #else
     this->jackServerSettings = storage.GetJackServerSettings();
+    this->jackConfiguration.AlsaInitialize(jackServerSettings);
 #endif
+
+    this->channelRouterSettings = storage.GetChannelRouterSettings();
+    pluginHost.OnConfigurationChanged(
+        jackConfiguration,
+        storage.GetChannelSelection());
+
 }
 
 
@@ -390,7 +397,7 @@ void PiPedalModel::Load()
     if (CrashGuard::HasCrashed())
     {
         // ignore the current preset, and load a blank pedalboard in order to avoid a potential plugin crash.
-        this->pedalboard = Pedalboard::MakeDefault();
+        this->pedalboard = Pedalboard::MakeDefault(Pedalboard::PedalboardType::MainPedalboard);
     }
     else
     {
@@ -1516,17 +1523,13 @@ void PiPedalModel::RestartAudio(bool useDummyAudioDriver)
             throw std::runtime_error("Audio configuration not valid.");
         }
 
-        auto channelSelection = this->storage.GetJackChannelSelection(jackConfiguration);
+        const auto & channelSelection = this->storage.GetChannelSelection();
 
-        if (!channelSelection.isValid())
-        {
-            throw std::runtime_error("Audio configuration not valid.");
-        }
         this->audioHost->Open(jackServerSettings, channelSelection);
 
         this->pluginHost.OnConfigurationChanged(jackConfiguration, channelSelection);
 
-        FireChannelSelectionChanged(-1);
+        FireChannelRouterSettingsChanged(-1);
         LoadCurrentPedalboard();
 
         this->UpdateRealtimeVuSubscriptions();
@@ -1622,33 +1625,18 @@ std::vector<AlsaSequencerPortSelection> PiPedalModel::GetAlsaSequencerPorts()
     return result;
 }
 
-void PiPedalModel::SetJackChannelSelection(int64_t clientId, const JackChannelSelection &channelSelection)
-{
-    {
-        std::lock_guard<std::recursive_mutex> lock(mutex); // copy atomically.
-        this->storage.SetJackChannelSelection(channelSelection);
 
-        this->pluginHost.OnConfigurationChanged(jackConfiguration, channelSelection);
-
-        CancelAudioRetry();
-    }
-
-    RestartAudio(); // no lock to avoid mutex deadlock when reader thread is sending notifications..
-
-    this->FireChannelSelectionChanged(clientId);
-}
-
-void PiPedalModel::FireChannelSelectionChanged(int64_t clientId)
+void PiPedalModel::FireChannelRouterSettingsChanged(int64_t clientId)
 {
     std::lock_guard<std::recursive_mutex> guard{mutex};
     {
         // take a snapshot incase a client unsusbscribes in the notification handler (in which case the mutex won't protect us)
-        JackChannelSelection channelSelection = storage.GetJackChannelSelection(this->jackConfiguration);
+        ChannelRouterSettings::ptr channelRouterSettings = this->channelRouterSettings;
 
         std::vector<IPiPedalModelSubscriber::ptr> t{subscribers.begin(), subscribers.end()};
         for (auto &subscriber : t)
         {
-            subscriber->OnChannelSelectionChanged(clientId, channelSelection);
+            subscriber->OnChannelRouterSettingsChanged(clientId, *channelRouterSettings);
         }
     }
 }
@@ -1765,7 +1753,7 @@ void PiPedalModel::UpdateRealtimeVuSubscriptions()
     for (int i = 0; i < activeVuSubscriptions.size(); ++i)
     {
         auto instanceId = activeVuSubscriptions[i].instanceid;
-        if (pedalboard.HasItem(instanceId) || instanceId == Pedalboard::INPUT_VOLUME_ID || instanceId == Pedalboard::OUTPUT_VOLUME_ID)
+        if (pedalboard.HasItem(instanceId) || instanceId == Pedalboard::START_CONTROL_ID || instanceId == Pedalboard::END_CONTROL_ID)
         {
             addedInstances.insert(activeVuSubscriptions[i].instanceid);
         }
@@ -3407,6 +3395,30 @@ int64_t PiPedalModel::CopyPresetsToBank(int64_t bankInstanceId, const std::vecto
     return lastAdded;
 
 }
+
+
+ChannelRouterSettings::ptr PiPedalModel::GetChannelRouterSettings()
+{
+    std::lock_guard<std::recursive_mutex> lock(mutex);
+    return this->channelRouterSettings;
+}
+
+void PiPedalModel::SetChannelRouterSettings(int64_t clientId,ChannelRouterSettings::ptr &settings)
+{
+    {
+        std::lock_guard<std::recursive_mutex> lock(mutex);
+        this->channelRouterSettings = settings;
+        this->storage.SetChannelRouterSettings(settings);
+        this->pluginHost.OnConfigurationChanged(jackConfiguration, *settings);
+        CancelAudioRetry();
+    }
+    RestartAudio(); // no lock to avoid mutex deadlock when reader thread is sending notifications..
+
+    this->FireChannelRouterSettingsChanged(clientId);
+
+}
+
+
 
 std::string PiPedalModel::Tone3000ThumbnailDirectory() 
 {