+++ src/Storage.hpp
@@ -32,6 +32,8 @@
 #include <map>
 #include "FilePropertyDirectoryTree.hpp"
 #include "AlsaSequencer.hpp"
+#include "ChannelRouterSettings.hpp"
+
 
 
 namespace pipedal {
@@ -96,7 +98,7 @@ private:
     void LoadCurrentBank();
     void SaveCurrentBank();
 
-    void LoadChannelSelection();
+    void LoadJackChannelSelection();
     void SaveChannelSelection();
 
     void LoadAlsaSequencerConfiguration();
@@ -115,6 +117,15 @@ private:
     WifiDirectConfigSettings wifiDirectConfigSettings;
 
     UserSettings userSettings;
+
+    ChannelRouterSettings::ptr channelRouterSettings;
+    ChannelSelection channelSelection;
+
+    void UpgradeChannelRouterSettings();
+    std::filesystem::path GetChannelRouterSettingsPath() const;
+    void SaveChannelRouterSettings(ChannelRouterSettings::ptr settings);
+    ChannelRouterSettings::ptr LoadChannelRouterSettings();
+
 public:
     Storage();
     void Initialize();
@@ -122,8 +133,8 @@ public:
 
     void SetDataRoot(const std::filesystem::path& path);
     void SetConfigRoot(const std::filesystem::path& path);
-    const std::filesystem::path&GetConfigRoot();
-    const std::filesystem::path&GetDataRoot();
+    const std::filesystem::path&GetConfigRoot() const;
+    const std::filesystem::path&GetDataRoot() const;
 
     const std::filesystem::path &GetPluginUploadDirectory() const;
 
@@ -203,6 +214,11 @@ public:
 
     void SetAlsaSequencerConfiguration(const AlsaSequencerConfiguration &alsaSequencerConfiguration);
     AlsaSequencerConfiguration GetAlsaSequencerConfiguration() const;
+
+    void SetChannelRouterSettings(ChannelRouterSettings::ptr settings);
+    ChannelRouterSettings::ptr GetChannelRouterSettings();
+
+    const ChannelSelection& GetChannelSelection() const;
 
 
     //std::string MapPropertyFileName(Lv2PluginInfo*pluginInfo, const std::string&path);